fill out auth functions, move password change to auth

+++ backend/src/controllers/profile.ts
@@ -1,13 +1,12 @@
 import { RequestHandler } from "express";
 import createHttpError from "http-errors";
 import prismaClient from "../db/client";
+import catchErrors from "../util/catchErrors";
+import { BAD_REQUEST, NOT_FOUND, UNAUTHORIZED } from "../config/constants";
+import bcrypt from "bcrypt";
 
-export const getAuthenticatedProfile: RequestHandler = async (
-  req,
-  res,
-  next
-) => {
-  try {
+export const getAuthenticatedProfile: RequestHandler = catchErrors(
+  async (req, res, next) => {
     const { profileId } = req.body;
 
     const profile = await prismaClient.profile.findUnique({
@@ -19,47 +18,60 @@ export const getAuthenticatedProfile: RequestHandler = async (
     if (!profile) throw createHttpError(404, "Profile not found");
 
     res.status(200).json(profile);
-  } catch (error) {
-    next(error);
   }
-};
+);
 
-export const changePassword: RequestHandler = async (req, res, next) => {
-  try {
-  } catch (error) {
-    next(error);
-  }
-};
+interface VerifyEmailBody {
+  profileId: number;
+  verificationCode: string;
+}
 
-export const verifyEmail: RequestHandler = async (req, res, next) => {
-  try {
-  } catch (error) {
-    next(error);
-  }
-};
+export const verifyEmail: RequestHandler<
+  unknown,
+  unknown,
+  VerifyEmailBody,
+  unknown
+> = catchErrors(async (req, res, next) => {
+  const { profileId, verificationCode } = req.body;
+});
 
-export const deleteProfile: RequestHandler = async (req, res, next) => {
-  try {
-    const { profileId } = req.body;
-    const sessions = await prismaClient.session.deleteMany({
-      where: { profileId },
-    });
-    if (!sessions.count) throw createHttpError(404, "Profile not found");
+interface DeleteProfilelBody {
+  profileId: number;
+  password: string;
+}
 
-    const profile = await prismaClient.profile.delete({
-      where: { id: profileId },
-    });
-    if (!profile) throw createHttpError(404, "Profile not found");
+export const deleteProfile: RequestHandler<
+  unknown,
+  unknown,
+  DeleteProfilelBody,
+  unknown
+> = catchErrors(async (req, res, next) => {
+  const { profileId, password } = req.body;
 
-    res.status(204);
-  } catch (error) {
-    next(error);
-  }
-};
+  if (!password) throw createHttpError(BAD_REQUEST, "password is required");
+
+  const profile = await prismaClient.profile.findUnique({
+    where: { id: profileId },
+  });
+  if (!profile) throw createHttpError(NOT_FOUND, "Profile not found");
+
+  const passwordMatch = await bcrypt.compare(password, profile.password);
+  if (!passwordMatch)
+    throw createHttpError(UNAUTHORIZED, "Invalid credentials");
+
+  await prismaClient.session.deleteMany({
+    where: { profileId },
+  });
+
+  await prismaClient.profile.delete({
+    where: { id: profileId },
+  });
+
+  res.status(204);
+});
 
 const profileApi = {
   getAuthenticatedProfile,
-  changePassword,
   verifyEmail,
   deleteProfile,
 };
